Axis Scaling Update for Limit Animator
Updated the limit animator to add some quality of life changes. Firstly,
the scaling of the y-axis was changed so that it tries to minimize its
range, making the graph easier to see. Secondly, I made the speed of the
animation dependent on the euclidean distance from the start point to
the end point so that the point doesn't move to fast on super steep
functions.